requirements.txt
streamlit==1.37.1
python-docx==1.1.2
PyPDF2==3.0.1
spacy==3.7.2
nltk==3.9.1
openai==1.43.0
textstat==0.7.4
rapidfuzz==3.9.6
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
httpx==0.27.2
